refactor(rust): Move note content and preimage building to util file
