+++ ironfish-rust/src/proofs/circuit/mint_asset.rs
@@ -1,13 +1,10 @@
-use std::slice;
-
 use bellman::{
     gadgets::{blake2s, boolean, multipack, num, Assignment},
     Circuit, ConstraintSystem,
 };
 use ff::PrimeField;
-use jubjub::ExtendedPoint;
 use zcash_primitives::{
-    constants::{self, GH_FIRST_BLOCK},
+    constants::{self},
     primitives::ProofGenerationKey,
 };
 use zcash_proofs::{circuit::ecc, constants::PROOF_GENERATION_KEY_GENERATOR};
@@ -18,7 +15,10 @@ use crate::{
         asset_type::AssetInfo, constants::ASSET_IDENTIFIER_PERSONALIZATION,
         sapling::ValueCommitment,
     },
-    proofs::circuit::sapling::{expose_value_commitment, slice_into_boolean_vec_le},
+    proofs::circuit::{
+        sapling::slice_into_boolean_vec_le,
+        util::{build_note_contents, hash_asset_info_to_preimage},
+    },
 };
 
 /// Info Needed:
@@ -51,42 +51,7 @@ impl Circuit<bls12_381::Scalar> for MintAsset {
         cs: &mut CS,
     ) -> Result<(), bellman::SynthesisError> {
         // Asset Commitment Contents
-        let mut identifier_commitment_contents = vec![];
-
-        // TODO: I wonder if we could hard-code this to minimize work?
-        // Not clear to me if the booleanizing is adding substantial time
-        // or if it's just a by-product of the hash taking longer due to
-        // more input. Also not clear if that has security implications
-        // by not witnessing the bits
-        let first_block_bits = slice_into_boolean_vec_le(
-            cs.namespace(|| "booleanize first block"),
-            Some(GH_FIRST_BLOCK),
-            64 * 8,
-        )?;
-
-        assert_eq!(first_block_bits.len(), 64 * 8);
-
-        let name_bits = slice_into_boolean_vec_le(
-            cs.namespace(|| "booleanize asset info name"),
-            self.asset_info.as_ref().and_then(|i| i.name().into()),
-            32 * 8,
-        )?;
-
-        let public_address_bits = slice_into_boolean_vec_le(
-            cs.namespace(|| "booleanize asset info public address"),
-            self.asset_info
-                .as_ref()
-                .and_then(|i| i.public_address_bytes().into()),
-            43 * 8,
-        )?;
-
-        let nonce_bits = slice_into_boolean_vec_le(
-            cs.namespace(|| "booleanize asset info nonce"),
-            self.asset_info
-                .as_ref()
-                .and_then(|i| slice::from_ref(i.nonce()).into()),
-            8,
-        )?;
+        let identifier_commitment_contents = hash_asset_info_to_preimage(cs, self.asset_info)?;
 
         // Public address validation
         // Prover witnesses ak (ensures that it's on the curve)
@@ -171,29 +136,14 @@ impl Circuit<bls12_381::Scalar> for MintAsset {
             .asset_info
             .as_ref()
             .and_then(|ai| Some(ai.asset_type()));
-        // Witness the asset type
-        // TODO: Does this properly verify that the spend note is the right asset generator?
-        // Could this be spoofed, or does this not matter? Need to consider what's public/private
-        // In other words: Does this verify that the actual note's asset generator is valid
-        // Or are we allowing this to be _any_ generator
-        let asset_generator = ecc::EdwardsPoint::witness(
-            cs.namespace(|| "asset_generator"),
-            asset_type
+
+        let public_address_bits = slice_into_boolean_vec_le(
+            cs.namespace(|| "booleanize asset info public address"),
+            self.asset_info
                 .as_ref()
-                .and_then(|at| at.asset_generator().into()),
+                .and_then(|i| i.public_address_bytes().into()),
+            43 * 8,
         )?;
-
-        let value_commitment_generator = ecc::EdwardsPoint::witness(
-            cs.namespace(|| "value commitment generator"),
-            asset_type
-                .as_ref()
-                .and_then(|at| ExtendedPoint::from(at.value_commitment_generator()).into()),
-        )?;
-
-        value_commitment_generator.assert_not_small_order(
-            cs.namespace(|| "value_commitment_generator not small order"),
-        )?;
-
         let calculated_pk_d_bits = pk_d.repr(cs.namespace(|| "representation of pk_d"))?;
         let asset_pk_d_bits = &public_address_bits[88..];
 
@@ -204,11 +154,6 @@ impl Circuit<bls12_381::Scalar> for MintAsset {
                 &calculated_pk_d_bits[i],
             )?;
         }
-
-        identifier_commitment_contents.extend(first_block_bits);
-        identifier_commitment_contents.extend(name_bits);
-        identifier_commitment_contents.extend(public_address_bits);
-        identifier_commitment_contents.extend(nonce_bits);
 
         let asset_identifier = blake2s::blake2s(
             cs.namespace(|| "blake2s(asset info)"),
@@ -317,50 +262,8 @@ impl Circuit<bls12_381::Scalar> for MintAsset {
             rt.inputize(cs.namespace(|| "anchor"))?;
         }
 
-        // Compute note contents:
-        // asset_generator, value (in big endian), g_d, pk_d
-        let mut note_contents = vec![];
-
-        // Place asset_generator in the note
-        note_contents
-            .extend(asset_generator.repr(cs.namespace(|| "representation of asset_generator"))?);
-
-        // Handle the value; we'll need it later for the
-        // dummy input check.
-        let mut value_num = num::Num::zero();
-        {
-            // Get the value in little-endian bit order
-            let value_bits = expose_value_commitment(
-                cs.namespace(|| "value commitment"),
-                value_commitment_generator,
-                self.value_commitment,
-            )?;
-
-            // Compute the note's value as a linear combination
-            // of the bits.
-            let mut coeff = bls12_381::Scalar::one();
-            for bit in &value_bits {
-                value_num = value_num.add_bool_with_coeff(CS::one(), bit, coeff);
-                coeff = coeff.double();
-            }
-
-            // Place the value in the note
-            note_contents.extend(value_bits);
-        }
-
-        // Place g_d in the note
-        note_contents.extend(g_d.repr(cs.namespace(|| "representation of g_d"))?);
-
-        // Place pk_d in the note
-        note_contents.extend(pk_d.repr(cs.namespace(|| "representation of pk_d"))?);
-
-        assert_eq!(
-            note_contents.len(),
-            256 + // asset_generator
-            64 + // value
-            256 + // g_d
-            256 // pk_d
-        );
+        let (note_contents, _) =
+            build_note_contents(cs, asset_type, self.value_commitment, g_d, pk_d)?;
 
         // Compute the hash of the note contents
         let mut cm = pedersen_hash::pedersen_hash(
